// src/lib/types.ts
export interface Product {
  id: string;
  name: string;
  sku?: string;
  quantity: number;
  minStockLevel: number;
  costPriceUSD: number;
  sellingPriceLYD: number;
  sellingPriceUSD?: number; // Optional secondary price
  description?: string;
  category?: string;
  createdAt?: string; // ISO String
  updatedAt?: string; // ISO String
}